@@ -30,7 +30,7 @@ aiofiles = "*"
 blinker = "*"
 click = "*"
 hypercorn = ">=0.11.2"
-importlib_metadata = { version = "*", python = "<3.8" }
+importlib_metadata = { version = "*", python = "<3.10" }
 itsdangerous = "*"
 jinja2 = "*"
 markupsafe = "*"
