@@ -3,4 +3,4 @@ from __future__ import annotations
 if __name__ == "__main__":
     from .cli import main
 
-    main(as_module=True)
+    main()
